Documenting the code

+++ src/main/scala/encoder/DNAEncoder.scala
@@ -1,11 +1,13 @@
 package encoder
 
+import org.apache.hadoop.fs.{FileSystem, Path}
+
 class DNAEncoder(subsequenceLength: Int) extends Serializable {
   private val DNABitsToSymbolSubstitutionTable: Array[Char] = Array[Char]('A', 'C', 'G', 'T')
   private val bitsByAlphabetSize = 2 // log10(DNABitsToSymbolSubstitutionTable.length)/log10(2.0)
   private val bitsMask = (1 << bitsByAlphabetSize) - 1 // 3 0x0000000000000011
 
-  private def getBitsFromChar(symbol: Char): Long = symbol match {
+  private def getBitsFromChar(symbol: Char): Int = symbol match {
     case 'A' | 'a' => 0
     case 'C' | 'c' => 1
     case 'G' | 'g' => 2
@@ -13,15 +15,15 @@ class DNAEncoder(subsequenceLength: Int) extends Serializable {
     case _ => 0
   }
 
-  def encodeSubsequenceToInteger(subSymbolList: String): Long = {
-    var encoded: Long = 0
+  def encodeSubsequenceToInteger(subSymbolList: String): Int = {
+    var encoded: Int = 0
     for (i <- 1 to subSymbolList.length) {
       encoded = encoded | (getBitsFromChar(subSymbolList.charAt(i - 1)) << ((subsequenceLength - i) * bitsByAlphabetSize))
     }
     encoded
   }
 
-  def decodeIntegerToString(encoded: Long, stringLength: Int = subsequenceLength): String = {
+  def decodeIntegerToString(encoded: Int, stringLength: Int = subsequenceLength): String = {
     val sb = new StringBuilder(stringLength)
     for (pos <- 0 until stringLength) {
       val posInInt = subsequenceLength - pos
@@ -31,6 +33,7 @@ class DNAEncoder(subsequenceLength: Int) extends Serializable {
     }
     sb.toString
   }
+
 }
 
 object DNAEncoder {
